refactor(commands): move detect/install logic into features Service layer
