+++ src/commands/installRuyi.ts
@@ -1,35 +1,20 @@
 // SPDX-License-Identifier: Apache-2.0
 /**
- * RuyiSDK VS Code Extension - Install via PyPI
+ * InstallCommand
  *
- * Command: ruyi.install
- * Workflow:
- *   1) Check if Python is available
- *   2) If Python is available → run `python -m pip install -U ruyi`
- *   3) Success → show "Ruyi installation completed."
- *   4) Failure → show error message
+ * VS Code command: `ruyi.install`
+ *
+ * Responsibilities:
+ * - Check platform support
+ * - Resolve Python via features/install service
+ * - Ask user for confirmation and show progress
+ * - Call features/install service to perform pip install and report result
  */
 
-import * as cp from 'child_process';
-import * as util from 'util';
 import * as vscode from 'vscode';
 
-import {LONG_CMD_TIMEOUT_MS, SHORT_CMD_TIMEOUT_MS} from '../common/constants';
-import {formatExecError, isSupportedPlatform, pythonCandidates} from '../common/utils';
-
-const execAsync = util.promisify(cp.exec);
-
-async function resolvePython(): Promise<string|null> {
-  for (const cmd of pythonCandidates()) {
-    try {
-      await execAsync(`${cmd} --version`, {timeout: SHORT_CMD_TIMEOUT_MS});
-      return cmd;
-    } catch {
-      // try next candidate
-    }
-  }
-  return null;
-}
+import {isSupportedPlatform} from '../common/utils';
+import {installViaPip, resolvePython} from '../features/install/InstallService';
 
 export default function registerInstallCommand(
     context: vscode.ExtensionContext) {
@@ -48,7 +33,10 @@ export default function registerInstallCommand(
     }
 
     const choice = await vscode.window.showInformationMessage(
-        'Python detected. Install/upgrade Ruyi via PyPI?', 'Install', 'Cancel');
+        'Python detected. Install/upgrade Ruyi via PyPI?',
+        'Install',
+        'Cancel',
+    );
     if (choice !== 'Install') return;
 
     await vscode.window.withProgress(
@@ -58,28 +46,25 @@ export default function registerInstallCommand(
           cancellable: false,
         },
         async () => {
-          try {
-            await execAsync(
-                `${py} -m pip install --user -U ruyi`,
-                {timeout: LONG_CMD_TIMEOUT_MS});
+          const result = await installViaPip(py);
 
-            try {
-              const {stdout} = await execAsync(`${py} -m ruyi --version`, {
-                timeout: SHORT_CMD_TIMEOUT_MS,
-              });
-              const version = stdout.trim();
-              vscode.window.showInformationMessage(
-                  `Ruyi installed: ${version}`);
-            } catch {
-              vscode.window.showWarningMessage(
-                  'Ruyi was installed, but the executable may not be in your PATH. ' +
-                  'If you used pip --user, add ~/.local/bin to PATH (e.g., export PATH="$HOME/.local/bin:$PATH").');
-            }
-          } catch (e: unknown) {
-            vscode.window.showErrorMessage(
-                `Failed to install Ruyi: ${formatExecError(e)}`);
+          if (result.errorMsg) {
+            vscode.window.showErrorMessage(result.errorMsg);
+            return;
           }
-        });
+          if (result.warnPath) {
+            vscode.window.showWarningMessage(
+                'Ruyi was installed, but the executable may not be in your PATH. ' +
+                    'If you used pip --user, add ~/.local/bin to PATH (e.g., export PATH="$HOME/.local/bin:$PATH").',
+            );
+            return;
+          }
+          if (result.version) {
+            vscode.window.showInformationMessage(
+                `Ruyi installed: ${result.version}`);
+          }
+        },
+    );
   });
 
   context.subscriptions.push(disposable);
